Added ImageCharacterViewModel
