+++ HotelDemo/View/HotelView.swift
@@ -11,14 +11,13 @@ struct HotelView: View {
 
     @EnvironmentObject var coordinator: Coordinator
     @StateObject private var viewModel = HotelViewModel()
-    @State private var currentIndex = 0
 
     var body: some View {
         ScrollView(showsIndicators: false) {
             VStack (spacing: 16) {
                 VStack {
                     VStack(spacing: 16) {
-                        imageCharacter()
+                        ImageCharacter(viewModel: viewModel.imagesCharacterVM)
                         basicInformation()
                     }
                     .padding()
@@ -86,52 +85,6 @@ struct HotelView_Previews: PreviewProvider {
 }
 
 extension HotelView {
-
-    private func imageCharacter() -> some View {
-        VStack {
-            if let images = viewModel.images {
-                TabView(selection: $currentIndex) {
-                    ForEach(0 ..< images.count, id: \.self) { index in
-                        Image(uiImage: images[index])
-                            .resizable()
-                            .scaledToFill()
-                            .frame(maxWidth: .infinity)
-                            .clipped()
-                            .background(.black)
-                    }
-                }
-                .frame(height: 257)
-                .tabViewStyle(PageTabViewStyle(indexDisplayMode: .never))
-                .cornerRadius(15)
-                .overlay {
-                    VStack {
-                        Spacer()
-                        HStack(spacing: 5) {
-                            ForEach(0..<images.count, id: \.self) { index in
-                                Circle()
-                                    .frame(width: 7, height: 7)
-                                    .foregroundColor(.black.opacity(viewModel.colorBarImageElement(index: index, currentIndex: currentIndex)))
-                                    .onTapGesture {
-                                        withAnimation {
-                                            currentIndex = index
-                                        }
-                                    }
-                            }
-                        }
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 5)
-                        .background(Color.white)
-                        .cornerRadius(5)
-                        .padding(8)
-                    }
-                }
-            } else {
-                ActivityIndicator()
-                    .frame(width: 100, height: 257)
-                    .foregroundColor(.gray)
-            }
-        }
-    }
 
     private func rating(rating: Int, ratingName: String) -> some View {
         HStack {
