+++ HotelDemo/ViewModel/HotelViewModel.swift
@@ -10,12 +10,12 @@ import UIKit
 
 class HotelViewModel: ObservableObject {
 
-    @Published var images: [UIImage]?
     @Published var hotel = Hotel.clearHotel
     let detailedInfo = [Detailed(name: "Удобства", imegeName: "Happy"),
                         Detailed(name: "Что включено", imegeName: "TickSquare"),
                         Detailed(name: "Что не включено", imegeName: "CloseSquare")]
-    var tagListVM = TagListViewModel()
+    let imagesCharacterVM = ImageCharacterViewModel()
+    let tagListVM = TagListViewModel()
 
     init() {
         getData()
@@ -27,7 +27,7 @@ class HotelViewModel: ObservableObject {
                 let hotel = try await NetworkServiceAA.shared.getData(dataset: Hotel.clearHotel)
                 DispatchQueue.main.async {
                     self.hotel = hotel
-                    self.getImages()
+                    self.imagesCharacterVM.getImages(urls: hotel.imageUrls)
                     self.tagListVM.setupTags(allTags: hotel.aboutTheHotel.peculiarities)
                 }
             } catch {
@@ -47,38 +47,6 @@ class HotelViewModel: ObservableObject {
         }
         let itog = "от " + priseString  + " ₽"
         return itog
-    }
-
-    func getImages() {
-        let imageUrls = hotel.imageUrls
-        imageUrls.forEach { url in
-            Task {
-                do {
-                    let image = try await NetworkServiceAA.shared.downloadImage(url: url)
-                    DispatchQueue.main.async {
-                        if let _ = self.images {
-                            self.images?.append(image)
-                        } else {
-                            self.images = [image]
-                        }
-                    }
-                } catch {
-                    print(error)
-                }
-            }
-        }
-    }
-    
-    func colorBarImageElement(index: Int, currentIndex: Int) -> Double {
-        if currentIndex == index {
-            return 1
-        } else {
-            if currentIndex > index {
-                return Double(index + 1) / Double(currentIndex * 5)
-            } else {
-                return Double(currentIndex + 1) / Double(index * 5)
-            }
-        }
     }
 
 }
